+++ src/components/index.js
@@ -10,7 +10,7 @@ import StInput from './input/index.vue';
 import StNotification from './notification/index.vue';
 import StNotificationsGroup from './notifications-group/index.vue';
 import StPagination from './pagination/index.vue';
-import StPopover from './popover/index.vue';
+import StPopper from './popper/index.vue';
 import StRadio from './radio/index.vue';
 import StRow from './row/index.vue';
 import StScrollbar from './scrollbar/index.vue';
@@ -28,7 +28,7 @@ const Components = {
   StNotification,
   StNotificationsGroup,
   StPagination,
-  StPopover,
+  StPopper,
   StRadio,
   StRow,
   StScrollbar,