+++ src/components/popper/types.ts
@@ -1,0 +1,24 @@
+export enum TriggerType {
+  click = 'click',
+  hover = 'hover',
+  focus = 'focus',
+  manual = 'manual',
+}
+
+export enum PopperPlacement {
+  autoStart = 'auto-start',
+  auto = 'auto',
+  autoEnd = 'auto-end',
+  topStart = 'top-start',
+  top = 'top',
+  topEnd = 'top-end',
+  rightStart = 'right-start',
+  right = 'right',
+  rightEnd = 'right-end',
+  bottomEnd = 'bottom-end',
+  bottom = 'bottom',
+  bottomStart = 'bottom-start',
+  leftEnd = 'left-end',
+  left = 'left',
+  leftStart = 'left-start',
+}